requirements.txt
openai
faiss-cpu
pandas
numpy
sentence-transformers
pyodbc
python-dotenv